Import the training dir from the ai-lab-recipes repository
This commit imports a specific commit ref (current `main`) of the
`training` directory of the github.com/containers/ai-lab-recipes
repository.

Updating the contents to a newer commit ref can be done by updating
the ref in the `Makefile` and then running `make update-training-dir`.

Note there is current a discrepency between the `README.md` and this
content. The `README.md` needs updates to reflect that all operations
are to be done from within the `training` directory.

Signed-off-by: Russell Bryant <rbryant@redhat.com>